Last update for the night
class bars are effectivly 100% done outside of monk bar, because fuck
that 5th bar from talents.
Tooltip is now sexy mode, HP text is not need to try some things
tomorrow
adjustments to healing frames i like them how they are now.
adjustments to boss frames, not sure if I like it as is yet.
Pet and Stance bars now match the flow of the UI 1px border and shit.